LPD-98986 Extract a shared asset list FDS props builder
