+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/props_transformer/utils/createScheduleDateFDSPropsTransformer.tsx
@@ -3,7 +3,6 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {IInternalRenderer, IItemsActions} from '@liferay/frontend-data-set-web';
 import React from 'react';
 
 import {ISearchAssetObjectEntry} from '../../../common/types/AssetType';
@@ -14,15 +13,10 @@ import {
 import {openCMSModal} from '../../../common/utils/openCMSModal';
 import ScheduleDateModalContent from '../../modal/ScheduleDateModalContent';
 import {triggerAssetBulkAction} from '../actions/triggerAssetBulkAction';
-import {getFileMimeTypeObjectDefinitionStickerValue} from './transformViewsItemProps';
-
-export interface AdditionalProps {
-	additionalAPIURLParameters?: string;
-	fileMimeTypeCssClasses?: Record<string, string>;
-	fileMimeTypeIcons?: Record<string, string>;
-	objectDefinitionCssClasses?: Record<string, string>;
-	objectDefinitionIcons?: Record<string, string>;
-}
+import {
+	AssetListFDSProps,
+	createAssetListFDSPropsBuilder,
+} from './createAssetListFDSPropsBuilder';
 
 interface ScheduleDateFDSConfig {
 	actionId: string;
@@ -43,36 +37,17 @@ interface ScheduleDateFDSConfig {
 export function createScheduleDateFDSPropsTransformer(
 	config: ScheduleDateFDSConfig
 ) {
-	function TitleRenderer({
-		itemData,
-		value,
-	}: {
-		itemData: ISearchAssetObjectEntry;
-		value?: string;
-	}) {
-		return (
-			<div className="d-flex flex-column">
-				<span>{value}</span>
-
-				<span className="font-weight-normal text-3">
-					{config.renderItemDate(itemData)}
-				</span>
-			</div>
-		);
-	}
+	const getAssetListFDSProps = createAssetListFDSPropsBuilder({
+		renderSubtitle: config.renderItemDate,
+		titleRendererName: config.titleRendererName,
+	});
 
 	return function ScheduleDateFDSPropsTransformer({
 		additionalProps,
 		itemsActions = [],
 		...otherProps
-	}: {
-		additionalProps: AdditionalProps;
-		apiURL?: string;
-		id?: string;
-		itemsActions?: IItemsActions[];
-	}) {
-		const {additionalAPIURLParameters, ...remainingAdditionalProps} =
-			additionalProps || {};
+	}: AssetListFDSProps) {
+		const {additionalAPIURLParameters} = additionalProps || {};
 
 		const bulkActionAPIURL =
 			additionalAPIURLParameters && otherProps.apiURL
@@ -115,20 +90,11 @@ export function createScheduleDateFDSPropsTransformer(
 		};
 
 		return {
-			...otherProps,
-			additionalAPIURLParameters,
-			additionalProps: remainingAdditionalProps,
-			customRenderers: {
-				listSection: [
-					{
-						component: TitleRenderer,
-						name: config.titleRendererName,
-						type: 'internal',
-					} as IInternalRenderer,
-				],
-			},
-			hideManagementBarInEmptyState: true,
-			itemsActions,
+			...getAssetListFDSProps({
+				additionalProps,
+				itemsActions,
+				...otherProps,
+			}),
 			onActionDropdownItemClick: ({
 				action,
 				event,
@@ -169,49 +135,6 @@ export function createScheduleDateFDSPropsTransformer(
 					direction: 'asc',
 					key: 'title',
 					label: Liferay.Language.get('title'),
-				},
-			],
-			views: [
-				{
-					contentRenderer: 'list',
-					default: true,
-					label: Liferay.Language.get('list'),
-					name: 'list',
-					schema: {
-						description: '',
-						sticker: 'sticker',
-						symbol: 'symbol',
-						title: 'title',
-						titleRendererName: config.titleRendererName,
-					},
-					setItemComponentProps: ({
-						item,
-						props,
-					}: {
-						item: ISearchAssetObjectEntry;
-						props: Record<string, unknown>;
-					}) => ({
-						...props,
-						item: {
-							...item,
-							sticker: {
-								className:
-									getFileMimeTypeObjectDefinitionStickerValue(
-										remainingAdditionalProps.fileMimeTypeCssClasses,
-										remainingAdditionalProps.objectDefinitionCssClasses ??
-											{},
-										item
-									),
-							},
-							symbol: getFileMimeTypeObjectDefinitionStickerValue(
-								remainingAdditionalProps.fileMimeTypeIcons,
-								remainingAdditionalProps.objectDefinitionIcons ??
-									{},
-								item
-							),
-						},
-					}),
-					thumbnail: 'list',
 				},
 			],
 		};
